# Use parallelism to reduce load time of tables