--- a/Project.toml
+++ b/Project.toml
@@ -6,14 +6,12 @@ version = "0.8.0"
 [deps]
 DataStructures = "864edb3b-99cc-5e75-8d2d-829cb0a9cfe8"
 HTTP = "cd3eb016-35fb-5094-929b-558a96fad6f3"
-ProgressMeter = "92933f4c-e287-5a05-a399-4b506db050ca"
 Transducers = "28d57a85-8fef-5791-bfe6-a80928e7c999"
 XMLDict = "228000da-037f-5747-90a9-8195ccbf91a5"
 
 [compat]
 DataStructures = "^0.16, 0.17"
 HTTP = "^0.8"
-ProgressMeter = "^1.2"
 XMLDict = "0.4"
 julia = "^1.3"