Before:
```julia-repl
julia> @benchmark  MortalityTables.tables()
BenchmarkTools.Trial:
  memory estimate:  2.44 GiB
  allocs estimate:  62552850
  --------------
  minimum time:     6.895 s (5.75% GC)
  median time:      6.895 s (5.75% GC)
  mean time:        6.895 s (5.75% GC)
  maximum time:     6.895 s (5.75% GC)
  --------------
  samples:          1
  evals/sample:     1
```

After:
```julia-repl
julia> @benchmark  MortalityTables.tables()
BenchmarkTools.Trial:
  memory estimate:  2.43 GiB
  allocs estimate:  62165872
  --------------
  minimum time:     2.709 s (24.99% GC)
  median time:      2.934 s (25.50% GC)
  mean time:        2.934 s (25.50% GC)
  maximum time:     3.160 s (25.94% GC)
  --------------
  samples:          2
  evals/sample:     1
```